Fix CI build for "linux" job (#9)
It seems the cimg/clojure:1.11.1-openjdk-17.0 Docker image used in the
"linux" job has been upgraded from Ubuntu 20.04 to Ubuntu 22.04. And
installing the "musl-tools" package and its dependencies already
places a file at /usr/bin/x86_64-linux-musl-gcc

That makes the last line of scripts/setup-musl to fail with the error

  ln: failed to create symbolic link '/usr/bin/x86_64-linux-musl-gcc': File exists

And aborts the "linux" job.

Thus we need to check if the file exists and, only if it doesn't, then
make the symbolic link.